pytorch_lightning/C-przechowalnia/requirements.txt
mlflow
ray[tune]
torch
pytorch-lightning
psycopg2
